Add time-dependent isotropic goal-oriented demo (#112)
@jwallwork23 - I found the cause of #107 : it was in how I was declaring
my `meshes` passed into the `GoalOrientedMeshSeq`. I was declaring
`meshes = [UnitSquareMesh[n,n, diagonal="left"] * n_meshes` which worked
when the `enrichment_method` was set to 'p' but produced an error when
accessing the `transfer` method between meshes when the
`enrichment_method` was set to 'h'. Presumably the mesh is a shallow
copy with the above expression.

Even though this turned into a user error, I found this while compiling
the MFC for burger's demo using goal-oriented isotropic adaptation which
may partially address #28 instead?

Creating a pull request here to get feedback on the additional demo.

---------

Co-authored-by: acse-ej321 <acse-ej321@github.com>
Co-authored-by: Joe Wallwork <22053413+jwallwork23@users.noreply.github.com>