Implement gradient computation w.r.t. initial conditions (#290)
Closes #284.

This PR sets up the ability to compute gradients of the QoI w.r.t. the
initial conditions. This will be extended to a wider range of controls
in #289.

Unit tests are added, which solve a simple R-space problem, where the
solution field is multiplied by a scalar at each timestep. Gradients of
steady-state, end-time, and time-integrated QoIs are computed and
compared against expected values.