Introduce `get_element` and `get_function_space` methods to `Field` class (#301)
See the comment chain in
https://github.com/mesh-adaptation/goalie/pull/296#discussion_r2030088537.
This branch merges into that one.

Instead of passing the `mesh` to `Field.__init__` and constructing a
finite element during initialisation, this PR introduces `get_element`
and `get_function_space` methods to `Field` that take the mesh as an
argument and return the element and fspace, respectively.

---------

Co-authored-by: Joe Wallwork <22053413+jwallwork23@users.noreply.github.com>