Drop `get_function_spaces` (#296)
Introduce get_element and get_function_space methods to Field c…
…lass (#301)

See the comment chain in
#296 (comment).
This branch merges into that one.

Instead of passing the `mesh` to `Field.__init__` and constructing a
finite element during initialisation, this PR introduces `get_element`
and `get_function_space` methods to `Field` that take the mesh as an
argument and return the element and fspace, respectively.

---------

Co-authored-by: Joe Wallwork <22053413+jwallwork23@users.noreply.github.com>

---------

Merges into #294.
Towards #189.

PR #294 introduces a new `Field` class, which holds metadata for fields,
including their finite elements. Given the meshes associated with a
`MeshSeq`, this information is sufficient to determine function spaces
on any of the meshes. As such, no longer need to pass a
`get_function_spaces` getter function to the `MeshSeq`.

---------

Co-authored-by: Davor Dundovic <33790330+ddundo@users.noreply.github.com>