Merge `develop` into `main` (#321)
I was wondering why the Goalie in the Docker image was still out of date
(and failing in https://github.com/mesh-adaptation/docs/pull/161) and
the reason why is that we've patched `develop` but not `main`.

I think this is a good opportunity to align them and then move to `main`
and `release` instead.

---------

Co-authored-by: Davor Dundovic <33790330+ddundo@users.noreply.github.com>